# Optimize useCustomProperties on native