Prevent unnecessary object creation and merging when dependencies haven't changed.

## Summary

| Task | Metric | Before | After | Difference | % Change |
|------|--------|--------|-------|------------|----------|
| **Nested divs** | Latency avg (ns) | 30,296,619.98 | 28,857,194.77 | -1,439,425.21 | -4.75% |
|  | Latency median (ns) | 29,549,489.50 | 28,850,770.50 | -698,719.00 | -2.36% |
|  | Throughput avg (ops/s) | 33 | 35 | +2 | +6.06% |
|  | Throughput median (ops/s) | 34 | 35 | +1 | +2.94% |
| **Sibling divs** | Latency avg (ns) | 29,540,653.28 | 29,049,781.09 | -490,872.19 | -1.66% |
|  | Latency median (ns) | 29,401,702.50 | 29,080,871.50 | -320,831.00 | -1.09% |
|  | Throughput avg (ops/s) | 34 | 34 | 0 | 0.00% |
|  | Throughput median (ops/s) | 34 | 34 | 0 | 0.00% |

## React Native Fantom benchmark results

### Before

| (index) | Task name                                                             | Latency average (ns)  | Latency median (ns)     | Throughput average (ops/s) | Throughput median (ops/s) | Samples |
| ------- | --------------------------------------------------------------------- | --------------------- | ----------------------- | -------------------------- | ------------------------- | ------- |
| 0       | 100 nested <html.div> with props and styles  | 30296619.98 ± 2.58% | 29549489.50 ± 3129.50 | 33 ± 1.90%               | 34                      | 64      |
| 1       | 100 sibling <html.div> with props and styles | 29540653.28 ± 0.54% | 29401702.50 ± 550.50  | 34 ± 0.52%              | 34                      | 64      |

### After

| (index) | Task name                                                             | Latency average (ns)  | Latency median (ns)      | Throughput average (ops/s) | Throughput median (ops/s) | Samples |
| ------- | --------------------------------------------------------------------- | --------------------- | ------------------------ | -------------------------- | ------------------------- | ------- |
| 0       | 100 nested <html.div> with props and styles  | 28857194.77 ± 0.41% | 28850770.50 ± 10270.50 | 35 ± 0.41%               | 35                      | 64      |
| 1       | 100 sibling <html.div> with props and styles | 29049781.09 ± 0.57% | 29080871.50 ± 7852.50  | 34 ± 0.56%               | 34                      | 64      |